# backend/OPENAPI.yaml
openapi: 3.0.3
info:
  title: Prescripto AI 2.0 API
  version: "1.0.0"
  x-aligned-to: "PRD v1 + System Architecture v1.1 (LOCKED) + PROJECT-SPEC v1.0"
  description: >
    Prescription document intelligence and medication-safety decision support.
    Canonical schema contract — see docs/API-CONTRACT.md for behavior not
    expressible here (idempotency, result-gating, review routing conditions).
servers:
  - url: /api/v1

paths:
  /auth/token:
    post:
      summary: Obtain a JWT access/refresh token pair
      operationId: createToken
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /auth/refresh:
    post:
      summary: Exchange a refresh token for a new access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: Expired, revoked, or blocklist unavailable (fails closed)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  /prescriptions:
    post:
      summary: Upload a prescription document
      operationId: createPrescription
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                patient_external_id: { type: string, nullable: true }
      responses:
        '202':
          description: Accepted — analysis enqueued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Idempotency key reused with a different file hash
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      summary: List prescriptions (caller-scoped, paginated)
      operationId: listPrescriptions
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PrescriptionList' }

  /prescriptions/{id}:
    get:
      summary: Get one document's metadata and a presigned image URL
      operationId: getPrescription
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Document metadata with a 60-second presigned image URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PrescriptionDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete a prescription document and all derived data
      operationId: deletePrescription
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '202':
          description: Deletion accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletionAccepted' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  /analyses/{id}:
    get:
      summary: Get analysis status and stage breakdown
      operationId: getAnalysisStatus
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Current status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalysisStatus' }
        '404': { $ref: '#/components/responses/NotFound' }

  /analyses/{id}/result:
    get:
      summary: Get full analysis result — only when COMPLETED or REQUIRES_REVIEW
      operationId: getAnalysisResult
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Full result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalysisResult' }
        '404':
          description: Not found, or not yet in a result-eligible state (QUEUED/PROCESSING/FAILED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  /analyses/{id}/review:
    post:
      summary: Submit reviewer decisions for an analysis
      operationId: submitReview
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewSubmission' }
      responses:
        '201':
          description: Review recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewRecord' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Analysis not in a reviewable state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  /health:
    get:
      summary: System health check
      operationId: getHealth
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthStatus' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest: { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Unauthorized: { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Forbidden: { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    NotFound: { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, example: ANALYSIS_NOT_READY }
            message: { type: string }
            request_id: { type: string, format: uuid }

    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, enum: [bearer] }
        expires_in: { type: integer, example: 900 }

    UploadAccepted:
      type: object
      required: [analysis_id, prescription_id]
      properties:
        analysis_id: { type: string, format: uuid }
        prescription_id: { type: string, format: uuid }

    DeletionAccepted:
      type: object
      required: [deletion_job_id]
      properties:
        deletion_job_id: { type: string, format: uuid }

    PrescriptionList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PrescriptionSummary' }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }

    PrescriptionSummary:
      type: object
      properties:
        prescription_id: { type: string, format: uuid }
        uploaded_at: { type: string, format: date-time }
        status: { type: string, enum: [UPLOAD_PENDING, UPLOADED, DELETION_REQUESTED, DELETION_IN_PROGRESS, DELETED] }
        latest_analysis_status: { type: string, nullable: true }

    PrescriptionDetail:
      type: object
      required: [prescription_id, status, uploaded_at, mime_type]
      properties:
        prescription_id: { type: string, format: uuid }
        status: { type: string, enum: [UPLOAD_PENDING, UPLOADED, DELETION_REQUESTED, DELETION_IN_PROGRESS, DELETED] }
        uploaded_at: { type: string, format: date-time }
        mime_type: { type: string }
        file_size_bytes: { type: integer }
        patient_ref: { type: string, nullable: true }
        image_url:
          type: string
          nullable: true
          description: >
            Presigned GET, 60-second TTL. A bearer capability — never cache, log, or persist it.
            Null when the document is deleted or upload is still pending.
        analyses:
          type: array
          items:
            type: object
            properties:
              analysis_id: { type: string, format: uuid }
              status: { type: string }
              created_at: { type: string, format: date-time }

    AnalysisStatus:
      type: object
      required: [analysis_id, status, stages]
      properties:
        analysis_id: { type: string, format: uuid }
        status:
          type: string
          enum: [QUEUED, PROCESSING, COMPLETED, REQUIRES_REVIEW, REVIEWING, REVIEWED_COMPLETE, REVIEWED_ESCALATED, FAILED]
        stages:
          type: array
          items:
            type: object
            properties:
              stage_name:
                type: string
                enum: [INGESTION, QUALITY_CHECK, TEXT_DETECTION, OCR_RECOGNITION,
                       STRUCTURED_EXTRACTION, MEDICATION_NORMALIZATION, SAFETY_SCREENING, REPORT_ASSEMBLY]
              status: { type: string, enum: [RUNNING, COMPLETED, FAILED, SKIPPED] }
              error_code: { type: string, nullable: true }

    FieldValue:
      type: object
      required: [value, state, confidence]
      properties:
        value: { type: string, nullable: true }
        state: { type: string, enum: [CLEAR, AMBIGUOUS, UNREADABLE, NOT_PRESENT] }
        confidence: { type: number, format: float, minimum: 0, maximum: 1 }

    MedicationLine:
      type: object
      description: >
        One extracted medication line. Field list matches SYSTEM-ARCHITECTURE.md §9
        prescription_medications (CR2-1 — includes unit and instructions).
      properties:
        line_index: { type: integer }
        name: { $ref: '#/components/schemas/FieldValue' }
        strength: { $ref: '#/components/schemas/FieldValue' }
        dose: { $ref: '#/components/schemas/FieldValue' }
        unit: { $ref: '#/components/schemas/FieldValue' }
        frequency: { $ref: '#/components/schemas/FieldValue' }
        route: { $ref: '#/components/schemas/FieldValue' }
        duration: { $ref: '#/components/schemas/FieldValue' }
        instructions: { $ref: '#/components/schemas/FieldValue' }
        # unit and instructions require the PROJECT-SPEC C-3 schema amendment before these serialize.
        resolved_medication_id: { type: string, format: uuid, nullable: true }
        resolution_status: { type: string, enum: [RESOLVED, UNRESOLVED, AMBIGUOUS] }

    RiskFindingOut:
      type: object
      required: [finding_id, finding_status, check_type, source_name, source_version, check_timestamp]
      properties:
        finding_id: { type: string, format: uuid }
        finding_status:
          type: string
          enum: [CONFIRMED_BY_SOURCE, POTENTIAL, INSUFFICIENT_EVIDENCE, NOT_EVALUATED, REQUIRES_REVIEW]
        check_type:
          type: string
          enum: [DUPLICATE_MEDICATION, EVIDENCE_LOOKUP, ADVERSE_EFFECT]
          description: V1 check types only. KNOWN_INTERACTION and DOSAGE_RANGE are deferred.
        medication_ids: { type: array, items: { type: string, format: uuid } }
        evidence_text: { type: string, nullable: true }
        source_name: { type: string }
        source_version: { type: string }
        confidence_score: { type: number, format: float, nullable: true }
        check_timestamp: { type: string, format: date-time }
        not_evaluated_reason: { type: string, nullable: true }

    NotEvaluatedItem:
      type: object
      properties:
        medication_ids: { type: array, items: { type: string, format: uuid } }
        reason:
          type: string
          enum: [SOURCE_UNAVAILABLE, PROVIDER_LACKS_CAPABILITY, UNRESOLVED_IDENTITY,
                 AMBIGUOUS_IDENTITY, SOURCE_NOT_COVERING, CHECK_TYPE_DEFERRED, LICENSE_MODE_EXCLUDED]

    AnalysisResult:
      type: object
      required: [analysis_id, status, analysis_timestamp, model_versions, medications, findings, not_evaluated, coverage_disclaimer]
      properties:
        analysis_id: { type: string, format: uuid }
        status: { type: string, enum: [COMPLETED, REQUIRES_REVIEW] }
        analysis_timestamp: { type: string, format: date-time }
        model_versions: { type: array, items: { type: string } }
        knowledge_snapshots: { type: array, items: { type: string } }
        medications:
          type: array
          items: { $ref: '#/components/schemas/MedicationLine' }
        findings:
          type: array
          items: { $ref: '#/components/schemas/RiskFindingOut' }
        not_evaluated:
          type: array
          items: { $ref: '#/components/schemas/NotEvaluatedItem' }
        coverage_disclaimer:
          type: string
          description: Mandatory (PRD §21.3) — cannot be omitted or empty.
        # NOTE: `explanation` is intentionally absent. LLM_ENABLED=false in V1 (ADR-09).
        # Re-adding it requires a PROJECT-SPEC update first — do not add it client-side.

    ReviewSubmission:
      type: object
      required: [decisions]
      properties:
        decisions:
          type: array
          items:
            type: object
            required: [finding_id, action]
            properties:
              finding_id: { type: string, format: uuid }
              action: { type: string, enum: [APPROVED, FLAGGED, ESCALATED] }
              notes: { type: string, nullable: true }

    ReviewRecord:
      type: object
      properties:
        review_id: { type: string, format: uuid }
        analysis_id: { type: string, format: uuid }
        reviewer_id: { type: string, format: uuid }
        submitted_at: { type: string, format: date-time }
        overall_status: { type: string, enum: [REVIEWED_COMPLETE, REVIEWED_ESCALATED] }

    HealthStatus:
      type: object
      properties:
        database: { type: string, enum: [ok, degraded, down] }
        object_storage: { type: string, enum: [ok, degraded, down] }
        worker_heartbeat: { type: string, enum: [ok, stale, down] }

security:
  - bearerAuth: []