# backend/prescripto/db/models/enums.py
"""
Canonical Enums for Prescripto AI 2.0.
Strictly matching docs/PROJECT-SPEC.md §3.
"""
from enum import Enum


class FieldState(str, Enum):
    CLEAR = "CLEAR"
    AMBIGUOUS = "AMBIGUOUS"
    UNREADABLE = "UNREADABLE"
    NOT_PRESENT = "NOT_PRESENT"


class ResolutionStatus(str, Enum):
    RESOLVED = "RESOLVED"
    UNRESOLVED = "UNRESOLVED"
    AMBIGUOUS = "AMBIGUOUS"


class FindingStatus(str, Enum):
    CONFIRMED_BY_SOURCE = "CONFIRMED_BY_SOURCE"
    POTENTIAL = "POTENTIAL"
    INSUFFICIENT_EVIDENCE = "INSUFFICIENT_EVIDENCE"
    NOT_EVALUATED = "NOT_EVALUATED"
    REQUIRES_REVIEW = "REQUIRES_REVIEW"


class SafetyCheckType(str, Enum):
    DUPLICATE_MEDICATION = "DUPLICATE_MEDICATION"
    EVIDENCE_LOOKUP = "EVIDENCE_LOOKUP"
    ADVERSE_EFFECT = "ADVERSE_EFFECT"
    # KNOWN_INTERACTION and DOSAGE_RANGE are deferred in V1


class AnalysisStatus(str, Enum):
    QUEUED = "QUEUED"
    PROCESSING = "PROCESSING"
    COMPLETED = "COMPLETED"
    REQUIRES_REVIEW = "REQUIRES_REVIEW"
    REVIEWING = "REVIEWING"
    REVIEWED_COMPLETE = "REVIEWED_COMPLETE"
    REVIEWED_ESCALATED = "REVIEWED_ESCALATED"
    FAILED = "FAILED"


class JobStatus(str, Enum):
    PENDING = "PENDING"
    CLAIMED = "CLAIMED"
    RUNNING = "RUNNING"
    SUCCEEDED = "SUCCEEDED"
    FAILED = "FAILED"
    DEAD = "DEAD"
    CANCELLED = "CANCELLED"


class StageStatus(str, Enum):
    RUNNING = "RUNNING"
    COMPLETED = "COMPLETED"
    FAILED = "FAILED"
    SKIPPED = "SKIPPED"


class DocumentStatus(str, Enum):
    UPLOAD_PENDING = "UPLOAD_PENDING"
    UPLOADED = "UPLOADED"
    DELETION_REQUESTED = "DELETION_REQUESTED"
    DELETION_IN_PROGRESS = "DELETION_IN_PROGRESS"
    DELETED = "DELETED"


class DeletionStatus(str, Enum):
    REQUESTED = "REQUESTED"
    DB_TOMBSTONED = "DB_TOMBSTONED"
    STORAGE_DELETING = "STORAGE_DELETING"
    VERIFYING = "VERIFYING"
    COMPLETE = "COMPLETE"
    PARTIAL_FAILURE = "PARTIAL_FAILURE"


class ReviewStatus(str, Enum):
    PENDING_ASSIGNMENT = "PENDING_ASSIGNMENT"
    ASSIGNED = "ASSIGNED"
    IN_PROGRESS = "IN_PROGRESS"
    SUBMITTED = "SUBMITTED"
    ESCALATED = "ESCALATED"


class ReviewAction(str, Enum):
    APPROVED = "APPROVED"
    FLAGGED = "FLAGGED"
    ESCALATED = "ESCALATED"


class Role(str, Enum):
    OPERATOR = "OPERATOR"
    REVIEWER = "REVIEWER"
    ADMIN = "ADMIN"


class LicenseMode(str, Enum):
    COMMERCIAL_PERMISSIVE = "COMMERCIAL_PERMISSIVE"
    RESEARCH_ONLY = "RESEARCH_ONLY"
    PUBLIC_DOMAIN = "PUBLIC_DOMAIN"


class PipelineStageName(str, Enum):
    INGESTION = "INGESTION"
    QUALITY_CHECK = "QUALITY_CHECK"
    TEXT_DETECTION = "TEXT_DETECTION"
    OCR_RECOGNITION = "OCR_RECOGNITION"
    STRUCTURED_EXTRACTION = "STRUCTURED_EXTRACTION"
    MEDICATION_NORMALIZATION = "MEDICATION_NORMALIZATION"
    SAFETY_SCREENING = "SAFETY_SCREENING"
    REPORT_ASSEMBLY = "REPORT_ASSEMBLY"